Moved tables to tables.dart
