+++ lib/data/db.dart
@@ -5,26 +5,9 @@ import 'package:drift/native.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as p;
 
+import './tables.dart';
+
 part 'db.g.dart';
-
-class Deck extends Table {
-  IntColumn get id => integer().autoIncrement()();
-  TextColumn get name => text().withLength(min: 1, max: 32)();
-}
-
-class Card extends Table {
-  IntColumn get id => integer().autoIncrement()();
-  IntColumn get deckId => integer().references(Deck, #id)();
-  TextColumn get front => text().withLength(min: 1, max: 512)();
-  TextColumn get back => text().withLength(min: 1, max: 512)();
-}
-
-class Review extends Table {
-  IntColumn get id => integer().autoIncrement()();
-  IntColumn get cardId => integer().references(Card, #id)();
-  DateTimeColumn get ts => dateTime()();
-  BoolColumn get correct => boolean()();
-}
 
 // this annotation tells drift to prepare a database class that uses both of the
 // tables we just defined. We'll see how to use that database class in a moment.
